+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 # ZIO-LMDB RELEASE NOTES
 
+## 1.6.0 - 2023-XX-YY
+
+- upgrade to lmdb-java 1.9.0
+- update dependencies
+- enhance collect / stream / streamWithKey (#19)
+  - forward or backward
+  - start after a given key
+- do not display logs during unit test execution
+
 ## 1.5.0 - 2023-09-24
 
 - add collection head, previous, next, last record operations (#18)